Move ErrorHandling to Utils Folder
