+++ Basecode.Services/Services/CharacterReferencesService.cs
@@ -8,6 +8,7 @@ using Basecode.Data.Dtos;
 using Basecode.Data.Interfaces;
 using Basecode.Data.Models;
 using Basecode.Services.Interfaces;
+using Basecode.Services.Utils;
 
 namespace Basecode.Services.Services
 {
